+++ lib/assets.php
@@ -49,6 +49,9 @@ add_action(
 		/* Theme styles and scripts */
 		$space_assets->enqueue( 'theme', 'main', [] );
 
+		/* Blocks animation */
+		$space_assets->enqueue( 'theme', 'blocksanimation', [] );
+
 		/* Floating header scripts */
 		if ( 'floating-header' === get_theme_mod( 'space_header_options', false ) ) {
 			$space_assets->enqueue( 'theme', 'floatingHeader', [] );