+++ src/scripts/editor.js
@@ -1,3 +1,0 @@
-/* Editor JS entry point */
-
-import './../styles/editor.scss';