+++ config/appearance.php
@@ -31,7 +31,7 @@ $space_accent_color = get_theme_mod(
 
 return [
 	'fonts-url'            => 'https://fonts.googleapis.com/css?family=Playfair+Display:400,700|Roboto:400,700|Roboto+Mono&display=optional',
-	'ionicons'             => 'https://unpkg.com/ionicons@4.2.2/dist/css/ionicons.min.css',
+	'ionicons'             => 'https://unpkg.com/ionicons@5.0.0/dist/ionicons.js',
 	'content-width'        => 1800,
 	'link-color'           => $space_link_color,
 	'default-colors'       => $space_default_colors,